+++ src/utils/random.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-pub fn get_random_num(nums: Vec<u8>) -> (u8, u8, u8) {
+pub fn get_random_num(nums: Vec<u32>) -> (u32, u32, u32) {
     let limit = nums[1];
     let init = nums[0];
